extend Playlist from `ArrayList` (#55)
Playlist was previously extended from `HashSet` to conviently remove
duplicate songs that may have been copied in when building the raw
`catalogue.json` file. This had the drawback of not preserving the
ordering and exact contents of a Playlist, which might want to have
duplicate songs or to be played in a specific order.